+++ docs/Changelog.md
@@ -3,5 +3,7 @@
 ## Next Version 0.1.8
 Added automatic installation of executors (JMeter, Grinder, Gatling).
   (See configuration options in ExecutionSettings.md)
+Added JunitXMLReporter. Reports test results in JunitXML format which may be used by Jenkins "JUnit test result report".
+  (See configuration options in Reporting.md)
 
 ## Prev Versions