+++ docs/Changelog.md
@@ -5,5 +5,6 @@ Added automatic installation of executors (JMeter, Grinder, Gatling).
   (See configuration options in ExecutionSettings.md)
 Added JunitXMLReporter. Reports test results in JunitXML format which may be used by Jenkins "JUnit test result report".
   (See configuration options in Reporting.md)
-
+Added JSON path assertion.
+  (See Assertions in ScenarioBuilding.md)
 ## Prev Versions