+++ docs/Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next Version 0.2.1
+  - Added pass/fail criteria report on console reporter.
 
 
 
@@ -9,5 +10,4 @@
   - Added JSON path assertion.
   - Added parameters for final_stats reporter
   - Added ability to generate query string based on parameters and request type.
-  - Added pass/fail criteria report on console reporter.